Consolidate assetOverrides into single file
